EPMLSTRTAI-113: EPMLSTRTAI-117: Suit's parameters validation were added to controller.

+++ Java/JDI/jdi-cucumber-test-generator/src/main/java/com/epam/test_generator/controllers/SuitController.java
@@ -20,8 +20,25 @@ public class SuitController {
     @Autowired
     SuitService suitService;
 
+    private boolean isPriorityValid(int priority) {
+    	return priority >= 1 && priority <= 5;
+	}
+
+	private boolean isNameValid(String name) {
+		return name != null && name.length() >= 1 && name.length() <= 255;
+	}
+
+	private boolean isDescriptionValid(String description) {
+		return description == null || description.length() <= 255;
+
+	}
+
+	private boolean isTagsValid(String tags) {
+		return tags == null || tags.length() <= 255;
+	}
+
 	@RequestMapping(value = "/")
-    public String getMainPage(){
+    public String getMainPage() {
         return "/WEB-INF/static/views/newSuits";
     }
 
@@ -33,37 +50,50 @@ public class SuitController {
 
     @RequestMapping(value = "/getSuit/{id}", method = RequestMethod.GET)
     @ResponseBody
-    public ResponseEntity<SuitDTO> getSuit(@PathVariable("id") long id){
+    public ResponseEntity<SuitDTO> getSuit(@PathVariable("id") long id) {
+    	SuitDTO suitDTO = suitService.getSuit(id);
 
-        return new ResponseEntity<>(suitService.getSuit(id), HttpStatus.OK);
-    }
+    	if (suitDTO.getId() == id) {
+			return new ResponseEntity<>(suitDTO, HttpStatus.OK);
+		}
 
-    @RequestMapping(value = "/getSuitByName/{name}", method = RequestMethod.GET)
-    @ResponseBody
-    public ResponseEntity<SuitDTO> getSuitByName(@PathVariable("name") String name){
-
-        return new ResponseEntity<>(suitService.getSuitByName(name),HttpStatus.OK);
+        return new ResponseEntity<>(suitDTO, HttpStatus.NOT_FOUND);
     }
 
     @RequestMapping(value="/updateSuit", method = RequestMethod.POST, consumes = "application/json")
-    public ResponseEntity<Void> editSuit(@RequestBody SuitDTO suit){
-        suitService.updateSuit(suit);
+    public ResponseEntity<Void> editSuit(@RequestBody SuitDTO suitDTO) {
+		if (isNameValid(suitDTO.getName()) && isPriorityValid(suitDTO.getPriority()) &&
+			isDescriptionValid(suitDTO.getDescription()) && isTagsValid(suitDTO.getTags())) {
+			suitService.updateSuit(suitDTO);
+
+			return new ResponseEntity<>(HttpStatus.OK);
+		}
       
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.UNPROCESSABLE_ENTITY);
     }
 
     @RequestMapping(value = "/removeSuit/{id}", method = RequestMethod.GET)
-    public ResponseEntity<Void> removeSuit(@PathVariable("id") long id){
-        suitService.removeSuit(id);
+    public ResponseEntity<Void> removeSuit(@PathVariable("id") long id) {
+		SuitDTO suitDTO = suitService.getSuit(id);
 
-        return new ResponseEntity<>(HttpStatus.OK);
+		if (suitDTO.getId() == id) {
+			suitService.removeSuit(id);
+
+			return new ResponseEntity<>(HttpStatus.OK);
+		}
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @RequestMapping(value="/addSuit", method = RequestMethod.POST, consumes = "application/json", produces = "application/json")
     @ResponseBody
-    public ResponseEntity<SuitDTO> addSuit(@RequestBody SuitDTO suit) {
+    public ResponseEntity<SuitDTO> addSuit(@RequestBody SuitDTO suitDTO) {
+		if (isNameValid(suitDTO.getName()) && isPriorityValid(suitDTO.getPriority()) &&
+			isDescriptionValid(suitDTO.getDescription()) && isTagsValid(suitDTO.getTags())) {
+			return new ResponseEntity<>(suitService.addSuit(suitDTO), HttpStatus.OK);
+		}
 
-        return new ResponseEntity<>(suitService.addSuit(suit), HttpStatus.OK);
+		return new ResponseEntity<>(suitDTO, HttpStatus.UNPROCESSABLE_ENTITY);
     }
 
     @ExceptionHandler(Exception.class)
